Started Project
I started the index and created my form. I created all the labels and gave each one either a select element with the proper options or a input with a type of radio, to give them the choice between Yes and No. I also added CSS for the background of the HTML file and to get the body somewhat going.