Form and  CSS
I added more to the form. I put  all the current form info I had into a fieldset and created a second fieldset for the customer information. I added CSS styles, giving the forms a flex display, validation coloring for the inputs, also gave the body more space and specified the widths of the fieldsets and making the fieldsets look overall better